move  _flatten and unflatten with temp sols
