+++ pennylane/optimize/qng.py
@@ -17,7 +17,7 @@ import pennylane as qml
 # pylint: disable=too-many-branches
 # pylint: disable=too-many-arguments
 from pennylane import numpy as pnp
-from pennylane.utils import _flatten, unflatten
+from pennylane.pytrees import flatten_np, unflatten_np
 
 from .gradient_descent import GradientDescentOptimizer
 
@@ -277,10 +277,10 @@ class QNGOptimizer(GradientDescentOptimizer):
         trained_index = 0
         for index, arg in enumerate(args):
             if getattr(arg, "requires_grad", False):
-                grad_flat = pnp.array(list(_flatten(grad[trained_index])))
+                grad_flat = pnp.array(list(flatten_np(grad[trained_index])))
                 # self.metric_tensor has already been reshaped to 2D, matching flat gradient.
                 update = pnp.linalg.pinv(mt[trained_index]) @ grad_flat
-                args_new[index] = arg - self.stepsize * unflatten(update, grad[trained_index])
+                args_new[index] = arg - self.stepsize * unflatten_np(update, grad[trained_index])
 
                 trained_index += 1
