+++ pennylane/pytrees/__init__.py
@@ -15,7 +15,16 @@
 An internal module for working with pytrees.
 """
 
-from .pytrees import PyTreeStructure, flatten, is_pytree, leaf, register_pytree, unflatten
+from .pytrees import (
+    PyTreeStructure,
+    flatten,
+    is_pytree,
+    leaf,
+    register_pytree,
+    unflatten,
+    flatten_np,
+    unflatten_np,
+)
 
 __all__ = [
     "PyTreeStructure",
@@ -24,4 +33,6 @@ __all__ = [
     "leaf",
     "register_pytree",
     "unflatten",
+    "flatten_np",
+    "unflatten_np",
 ]
