+++ tests/test_utils.py
@@ -38,55 +38,6 @@ test_shapes = [
 ]
 
 
-class TestFlatten:
-    """Tests the flatten and unflatten functions"""
-
-    @pytest.mark.parametrize("shape", test_shapes)
-    def test_flatten(self, shape):
-        """Tests that _flatten successfully flattens multidimensional arrays."""
-
-        reshaped = np.reshape(flat_dummy_array, shape)
-        flattened = np.array(list(pu._flatten(reshaped)))
-
-        assert flattened.shape == flat_dummy_array.shape
-        assert np.array_equal(flattened, flat_dummy_array)
-
-    @pytest.mark.parametrize("shape", test_shapes)
-    def test_unflatten(self, shape):
-        """Tests that _unflatten successfully unflattens multidimensional arrays."""
-
-        reshaped = np.reshape(flat_dummy_array, shape)
-        unflattened = np.array(list(pu.unflatten(flat_dummy_array, reshaped)))
-
-        assert unflattened.shape == reshaped.shape
-        assert np.array_equal(unflattened, reshaped)
-
-    def test_unflatten_error_unsupported_model(self):
-        """Tests that unflatten raises an error if the given model is not supported"""
-
-        with pytest.raises(TypeError, match="Unsupported type in the model"):
-            model = lambda x: x  # not a valid model for unflatten
-            pu.unflatten(flat_dummy_array, model)
-
-    def test_unflatten_error_too_many_elements(self):
-        """Tests that unflatten raises an error if the given iterable has
-        more elements than the model"""
-
-        reshaped = np.reshape(flat_dummy_array, (16, 2, 2))
-
-        with pytest.raises(ValueError, match="Flattened iterable has more elements than the model"):
-            pu.unflatten(np.concatenate([flat_dummy_array, flat_dummy_array]), reshaped)
-
-    def test_flatten_wires(self):
-        """Tests flattening a Wires object."""
-        wires = qml.wires.Wires([3, 4])
-        wires_int = [3, 4]
-
-        wires = qml.utils._flatten(wires)
-        for i, wire in enumerate(wires):
-            assert wires_int[i] == wire
-
-
 class TestArgumentHelpers:
     """Tests for auxiliary functions to help with parsing
     Python function arguments"""
